+++ CSM Security Database Core/Abstractions/Bases/SecurityNamedEntityBase.cs
@@ -1,14 +1,6 @@
 ﻿using System.Text.Json.Serialization;
 
-using CSM_Database_Core.Core.Attributes;
-using CSM_Database_Core.Entities.Abstractions.Bases;
-
-using CSM_Security_Database_Core.Entities;
-
-using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Metadata.Builders;
-
-using EntityState = CSM_Security_Database_Core.Entities.EntityState;
+using CSM_System_Database_Core.Abstractions.Bases;
 
 namespace CSM_Security_Database_Core.Abstractions.Bases;
 
@@ -19,37 +11,10 @@ namespace CSM_Security_Database_Core.Abstractions.Bases;
 ///     Usage must be exclusively for [CSM Database Security] entities.
 /// </remarks>
 public abstract class SecurityNamedEntityBase
-    : NamedEntityBase {
+    : StateSystemNamedEntityBase {
 
     /// <inheritdoc/>
     [JsonIgnore]
     public override Type Database { get; init; } = typeof(SecurityDatabase);
 
-    /// <inheritdoc/>
-    [EntityRelation]
-    public EntityState State { get; set; } = default!;
-
-    /// <inheritdoc/>
-    protected override void DesignEntity(EntityTypeBuilder etBuilder) {
-        base.DesignEntity(etBuilder);
-
-        const string shadowProperty = $"{nameof(State)}Shadow";
-
-        etBuilder
-            .Property<long>(shadowProperty)
-            .HasColumnName(nameof(State))
-            .HasColumnType("bigint")
-            .IsRequired();
-
-        etBuilder
-            .HasOne(typeof(EntityState), nameof(State))
-            .WithMany()
-            .HasForeignKey(shadowProperty)
-            .OnDelete(DeleteBehavior.Restrict)
-            .IsRequired();
-
-        etBuilder
-            .Navigation(nameof(State))
-            .AutoInclude();
-    }
 }