+++ CSM Security Database Core/Migrations/20260911234718_8_including-state-property.Designer.cs
@@ -12,8 +12,8 @@ using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 namespace CSM_Security.Migrations
 {
     [DbContext(typeof(SecurityDatabase))]
-    [Migration("20260815123532_8_incluiding_entities_states")]
-    partial class _8_incluiding_entities_states
+    [Migration("20260911234718_8_including-state-property")]
+    partial class _8_includingstateproperty
     {
         /// <inheritdoc />
         protected override void BuildTargetModel(ModelBuilder modelBuilder)
@@ -41,41 +41,9 @@ namespace CSM_Security.Migrations
                         .HasMaxLength(100)
                         .HasColumnType("nvarchar(100)");
 
-                    b.Property<long>("StateShadow")
-                        .HasColumnType("bigint")
-                        .HasColumnName("State");
-
-                    b.Property<DateTime>("Timestamp")
-                        .ValueGeneratedOnAdd()
-                        .HasColumnType("datetime2(7)")
-                        .HasDefaultValueSql("GETUTCDATE()");
-
-                    b.HasKey("Id");
-
-                    b.HasIndex("Name")
-                        .IsUnique();
-
-                    b.HasIndex("StateShadow");
-
-                    b.ToTable("Actions");
-                });
-
-            modelBuilder.Entity("CSM_Security_Database_Core.Entities.EntityState", b =>
-                {
-                    b.Property<long>("Id")
-                        .ValueGeneratedOnAdd()
+                    b.Property<long>("State")
                         .HasColumnType("bigint");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"));
-
-                    b.Property<string>("Description")
-                        .HasColumnType("nvarchar(max)");
-
-                    b.Property<string>("Name")
-                        .IsRequired()
-                        .HasMaxLength(100)
-                        .HasColumnType("nvarchar(100)");
-
                     b.Property<DateTime>("Timestamp")
                         .ValueGeneratedOnAdd()
                         .HasColumnType("datetime2(7)")
@@ -86,7 +54,7 @@ namespace CSM_Security.Migrations
                     b.HasIndex("Name")
                         .IsUnique();
 
-                    b.ToTable("EntityStates");
+                    b.ToTable("Actions");
                 });
 
             modelBuilder.Entity("CSM_Security_Database_Core.Entities.Feature", b =>
@@ -105,9 +73,8 @@ namespace CSM_Security.Migrations
                         .HasMaxLength(100)
                         .HasColumnType("nvarchar(100)");
 
-                    b.Property<long>("StateShadow")
-                        .HasColumnType("bigint")
-                        .HasColumnName("State");
+                    b.Property<long>("State")
+                        .HasColumnType("bigint");
 
                     b.Property<DateTime>("Timestamp")
                         .ValueGeneratedOnAdd()
@@ -118,8 +85,6 @@ namespace CSM_Security.Migrations
 
                     b.HasIndex("Name")
                         .IsUnique();
-
-                    b.HasIndex("StateShadow");
 
                     b.ToTable("Features");
                 });
@@ -152,9 +117,8 @@ namespace CSM_Security.Migrations
                         .HasColumnType("bigint")
                         .HasColumnName("Solution");
 
-                    b.Property<long>("StateShadow")
-                        .HasColumnType("bigint")
-                        .HasColumnName("State");
+                    b.Property<long>("State")
+                        .HasColumnType("bigint");
 
                     b.Property<DateTime>("Timestamp")
                         .ValueGeneratedOnAdd()
@@ -169,8 +133,6 @@ namespace CSM_Security.Migrations
                         .IsUnique();
 
                     b.HasIndex("SolutionShadow");
-
-                    b.HasIndex("StateShadow");
 
                     b.HasIndex("ActionShadow", "SolutionShadow", "FeatureShadow");
 
@@ -193,9 +155,8 @@ namespace CSM_Security.Migrations
                         .HasMaxLength(100)
                         .HasColumnType("nvarchar(100)");
 
-                    b.Property<long>("StateShadow")
-                        .HasColumnType("bigint")
-                        .HasColumnName("State");
+                    b.Property<long>("State")
+                        .HasColumnType("bigint");
 
                     b.Property<DateTime>("Timestamp")
                         .ValueGeneratedOnAdd()
@@ -206,8 +167,6 @@ namespace CSM_Security.Migrations
 
                     b.HasIndex("Name")
                         .IsUnique();
-
-                    b.HasIndex("StateShadow");
 
                     b.ToTable("Profiles");
                 });
@@ -234,9 +193,8 @@ namespace CSM_Security.Migrations
                         .HasColumnType("nchar(5)")
                         .IsFixedLength();
 
-                    b.Property<long>("StateShadow")
-                        .HasColumnType("bigint")
-                        .HasColumnName("State");
+                    b.Property<long>("State")
+                        .HasColumnType("bigint");
 
                     b.Property<DateTime>("Timestamp")
                         .ValueGeneratedOnAdd()
@@ -250,8 +208,6 @@ namespace CSM_Security.Migrations
 
                     b.HasIndex("Sign")
                         .IsUnique();
-
-                    b.HasIndex("StateShadow");
 
                     b.ToTable("Solutions");
                 });
@@ -271,9 +227,8 @@ namespace CSM_Security.Migrations
                         .IsRequired()
                         .HasColumnType("varbinary(max)");
 
-                    b.Property<long>("StateShadow")
-                        .HasColumnType("bigint")
-                        .HasColumnName("State");
+                    b.Property<long>("State")
+                        .HasColumnType("bigint");
 
                     b.Property<DateTime>("Timestamp")
                         .ValueGeneratedOnAdd()
@@ -293,8 +248,6 @@ namespace CSM_Security.Migrations
                         .HasColumnType("nvarchar(50)");
 
                     b.HasKey("Id");
-
-                    b.HasIndex("StateShadow");
 
                     b.HasIndex("UserInfoShadow")
                         .IsUnique();
@@ -333,9 +286,8 @@ namespace CSM_Security.Migrations
                         .HasMaxLength(14)
                         .HasColumnType("nvarchar(14)");
 
-                    b.Property<long>("StateShadow")
-                        .HasColumnType("bigint")
-                        .HasColumnName("State");
+                    b.Property<long>("State")
+                        .HasColumnType("bigint");
 
                     b.Property<DateTime>("Timestamp")
                         .ValueGeneratedOnAdd()
@@ -349,8 +301,6 @@ namespace CSM_Security.Migrations
 
                     b.HasIndex("Phone")
                         .IsUnique();
-
-                    b.HasIndex("StateShadow");
 
                     b.ToTable("UserInfos");
                 });
@@ -371,9 +321,8 @@ namespace CSM_Security.Migrations
                         .HasMaxLength(100)
                         .HasColumnType("nvarchar(100)");
 
-                    b.Property<long>("StateShadow")
-                        .HasColumnType("bigint")
-                        .HasColumnName("State");
+                    b.Property<long>("State")
+                        .HasColumnType("bigint");
 
                     b.Property<DateTime>("Timestamp")
                         .ValueGeneratedOnAdd()
@@ -387,8 +336,6 @@ namespace CSM_Security.Migrations
 
                     b.HasIndex("Name")
                         .IsUnique();
-
-                    b.HasIndex("StateShadow");
 
                     b.ToTable("Vendors");
                 });
@@ -453,28 +400,6 @@ namespace CSM_Security.Migrations
                     b.ToTable("Users_Vendors");
                 });
 
-            modelBuilder.Entity("CSM_Security_Database_Core.Entities.Action", b =>
-                {
-                    b.HasOne("CSM_Security_Database_Core.Entities.EntityState", "State")
-                        .WithMany()
-                        .HasForeignKey("StateShadow")
-                        .OnDelete(DeleteBehavior.Restrict)
-                        .IsRequired();
-
-                    b.Navigation("State");
-                });
-
-            modelBuilder.Entity("CSM_Security_Database_Core.Entities.Feature", b =>
-                {
-                    b.HasOne("CSM_Security_Database_Core.Entities.EntityState", "State")
-                        .WithMany()
-                        .HasForeignKey("StateShadow")
-                        .OnDelete(DeleteBehavior.Restrict)
-                        .IsRequired();
-
-                    b.Navigation("State");
-                });
-
             modelBuilder.Entity("CSM_Security_Database_Core.Entities.Permit", b =>
                 {
                     b.HasOne("CSM_Security_Database_Core.Entities.Action", "Action")
@@ -495,82 +420,22 @@ namespace CSM_Security.Migrations
                         .OnDelete(DeleteBehavior.Restrict)
                         .IsRequired();
 
-                    b.HasOne("CSM_Security_Database_Core.Entities.EntityState", "State")
-                        .WithMany()
-                        .HasForeignKey("StateShadow")
-                        .OnDelete(DeleteBehavior.Restrict)
-                        .IsRequired();
-
                     b.Navigation("Action");
 
                     b.Navigation("Feature");
 
                     b.Navigation("Solution");
-
-                    b.Navigation("State");
-                });
-
-            modelBuilder.Entity("CSM_Security_Database_Core.Entities.Profile", b =>
-                {
-                    b.HasOne("CSM_Security_Database_Core.Entities.EntityState", "State")
-                        .WithMany()
-                        .HasForeignKey("StateShadow")
-                        .OnDelete(DeleteBehavior.Restrict)
-                        .IsRequired();
-
-                    b.Navigation("State");
-                });
-
-            modelBuilder.Entity("CSM_Security_Database_Core.Entities.Solution", b =>
-                {
-                    b.HasOne("CSM_Security_Database_Core.Entities.EntityState", "State")
-                        .WithMany()
-                        .HasForeignKey("StateShadow")
-                        .OnDelete(DeleteBehavior.Restrict)
-                        .IsRequired();
-
-                    b.Navigation("State");
                 });
 
             modelBuilder.Entity("CSM_Security_Database_Core.Entities.User", b =>
                 {
-                    b.HasOne("CSM_Security_Database_Core.Entities.EntityState", "State")
-                        .WithMany()
-                        .HasForeignKey("StateShadow")
-                        .OnDelete(DeleteBehavior.Restrict)
-                        .IsRequired();
-
                     b.HasOne("CSM_Security_Database_Core.Entities.UserInfo", "UserInfo")
                         .WithOne()
                         .HasForeignKey("CSM_Security_Database_Core.Entities.User", "UserInfoShadow")
                         .OnDelete(DeleteBehavior.Cascade)
                         .IsRequired();
 
-                    b.Navigation("State");
-
                     b.Navigation("UserInfo");
-                });
-
-            modelBuilder.Entity("CSM_Security_Database_Core.Entities.UserInfo", b =>
-                {
-                    b.HasOne("CSM_Security_Database_Core.Entities.EntityState", "State")
-                        .WithMany()
-                        .HasForeignKey("StateShadow")
-                        .OnDelete(DeleteBehavior.Restrict)
-                        .IsRequired();
-
-                    b.Navigation("State");
-                });
-
-            modelBuilder.Entity("CSM_Security_Database_Core.Entities.Vendor", b =>
-                {
-                    b.HasOne("CSM_Security_Database_Core.Entities.EntityState", "State")
-                        .WithMany()
-                        .HasForeignKey("StateShadow")
-                        .OnDelete(DeleteBehavior.Restrict)
-                        .IsRequired();
-
-                    b.Navigation("State");
                 });
 
             modelBuilder.Entity("Profiles_Permits", b =>