+++ CSM Security Database Core/CHANGELOG.md
@@ -4,13 +4,14 @@
 
 ### Patch
 
-- Added [Entity State] entity to track the state of entities in the database.
+- Added [CSM.System.Database.Core] state property bases to security entities to track their states in the database.
 
 #### Dependencies
 
 | Package                                 | Previous Version | New Version     |
 |:----------------------------------------|:----------------:|:---------------:|
 | CSM.Database.Core                       | 6.1.0            | 7.0.0           |
+| CSM.System.Database.Core                | -.-.-            | 1.0.0           |
 
 ## [3.1.0] - 24.06.2026